Split main.rs into modular components
- Extract pretty printing utilities into utils/pretty_print.rs
- Separate CLI commands into individual modules (ssh_keys.rs, pgp_keys.rs, known_hosts.rs)
- Create proper module structure with mod.rs files for better organization
- Maintain all existing functionality while improving code maintainability and readability

Note: This commit was mostly authored by Github Copilot Agent using Claude Sonnet 3.7
