+++ cli/src/commands/mod.rs
@@ -1,0 +1,8 @@
+pub mod known_hosts;
+pub mod pgp_keys;
+pub mod ssh_keys;
+
+// Re-export the main command functions for easier imports
+pub use known_hosts::fetch_known_hosts;
+pub use pgp_keys::fetch_pgp_keys;
+pub use ssh_keys::fetch_ssh_keys;
