+++ cli/src/utils/mod.rs
@@ -1,0 +1,3 @@
+pub mod pretty_print;
+
+pub use pretty_print::*;
